// vue/src/store/generated/stateset/stateset-blockchain/stateset.inovice.v1beta1/module/types/stateset/invoice/v1beta1/tx.d.ts
export interface Invoice {
    creator: string;
    invoiceId: string;
    invoiceNumber: string;
    invoiceName: string;
    billingReason: string;
    amountDue: string;
    amountPaid: string;
    amountRemaining: string;
    subtotal: string;
    total: string;
    party: string;
    counterparty: string;
    factor: string;
    dueDate: string;
    periodStartDate: string;
    periodEndDate: string;
}
// ...
export declare const Invoice: {
    encode(message: Invoice, writer?: Writer): Writer;
    decode(input: Reader | Uint8Array, length?: number): Invoice;
    fromJSON(object: any): Invoice;
    toJSON(message: Invoice): unknown;
    fromPartial(object: DeepPartial<Invoice>): Invoice;
};
// ...
declare type Builtin = Date | Function | Uint8Array | string | number | undefined;
export declare type DeepPartial<T> = T extends Builtin ? T : T extends Array<infer U> ? Array<DeepPartial<U>> : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>> : T extends {} ? {
    [K in keyof T]?: DeepPartial<T[K]>;
} : Partial<T>;